Rename the SetPasswordPage => SetPasswordForm, change its styles to match that of the sign in page.
