+++ src/pages/SetPasswordForm.js
@@ -1,0 +1,131 @@
+import React, {Component} from 'react';
+import {
+    Text,
+    TextInput,
+    View,
+} from 'react-native';
+import PropTypes from 'prop-types';
+import {withOnyx} from 'react-native-onyx';
+import _ from 'underscore';
+import lodashGet from 'lodash/get';
+import styles from '../styles/styles';
+import {setPassword} from '../libs/actions/Session';
+import ONYXKEYS from '../ONYXKEYS';
+import ButtonWithLoader from '../components/ButtonWithLoader';
+import themeColors from '../styles/themes/default';
+import SignInPageLayout from './signin/SignInPageLayout';
+
+const propTypes = {
+    /* Onyx Props */
+
+    // The details about the account that the user is signing in with
+    account: PropTypes.shape({
+        // An error message to display to the user
+        error: PropTypes.string,
+
+        // Whether or not a sign on form is loading (being submitted)
+        loading: PropTypes.bool,
+    }),
+
+    // The credentials of the logged in person
+    credentials: PropTypes.shape({
+        // The email the user logged in with
+        login: PropTypes.string,
+
+        // The password used to log in the user
+        password: PropTypes.string,
+    }),
+
+    route: PropTypes.shape({
+        params: PropTypes.shape({
+            validateCode: PropTypes.string,
+        }),
+    }),
+};
+
+const defaultProps = {
+    account: {},
+    credentials: {},
+    route: {
+        params: {},
+    },
+};
+
+class SetPasswordForm extends Component {
+    constructor(props) {
+        super(props);
+
+        this.submitForm = this.submitForm.bind(this);
+
+        this.state = {
+            password: '',
+            formError: null,
+        };
+    }
+
+    /**
+     * Validate the form and then submit it
+     */
+    submitForm() {
+        if (!this.state.password.trim()) {
+            this.setState({
+                formError: 'Password cannot be blank',
+            });
+            return;
+        }
+
+        this.setState({
+            formError: null,
+        });
+        setPassword(this.state.password, lodashGet(this.props.route, 'params.validateCode', ''));
+    }
+
+    render() {
+        return (
+            <SignInPageLayout>
+                <View style={[styles.mb4]}>
+                    <Text style={[styles.formLabel]}>Enter a password:</Text>
+                    <TextInput
+                        style={[styles.textInput]}
+                        value={this.state.login}
+                        autoCompleteType="email"
+                        textContentType="username"
+                        onChangeText={text => this.setState({login: text})}
+                        onSubmitEditing={this.validateAndSubmitForm}
+                        autoCapitalize="none"
+                        placeholder="Phone or Email"
+                        placeholderTextColor={themeColors.placeholderText}
+                        autofocus
+                    />
+                </View>
+                <View>
+                    <ButtonWithLoader
+                        text="Set Password"
+                        isLoading={this.props.account.loading}
+                        onClick={this.validateAndSubmitForm}
+                    />
+                </View>
+
+                {this.state.formError && (
+                    <Text style={[styles.formError]}>
+                        {this.state.formError}
+                    </Text>
+                )}
+
+                {!_.isEmpty(this.props.account.error) && (
+                    <Text style={[styles.formError]}>
+                        {this.props.account.error}
+                    </Text>
+                )}
+            </SignInPageLayout>
+        );
+    }
+}
+
+SetPasswordForm.propTypes = propTypes;
+SetPasswordForm.defaultProps = defaultProps;
+
+export default withOnyx({
+    credentials: {key: ONYXKEYS.CREDENTIALS},
+    account: {key: ONYXKEYS.ACCOUNT},
+})(SetPasswordForm);
